+++ Sources/StreamChatUI/ChatThread/ChatThreadVC.swift
@@ -29,7 +29,8 @@ open class ChatThreadVC: _ViewController,
     /// Component responsible for setting the correct offset when keyboard frame is changed
     open lazy var keyboardHandler: KeyboardHandler = ComposerKeyboardHandler(
         composerParentVC: self,
-        composerBottomConstraint: messageComposerBottomConstraint
+        composerBottomConstraint: messageComposerBottomConstraint,
+        messageListVC: messageListVC
     )
 
     /// User search controller passed directly to the composer
@@ -57,14 +58,6 @@ open class ChatThreadVC: _ViewController,
 
     override open func setUp() {
         super.setUp()
-
-        let notificationCenter = NotificationCenter.default
-        notificationCenter.addObserver(
-            self,
-            selector: #selector(appMovedToForeground),
-            name: UIApplication.willEnterForegroundNotification,
-            object: nil
-        )
 
         messageListVC.delegate = self
         messageListVC.dataSource = self
@@ -338,13 +331,6 @@ open class ChatThreadVC: _ViewController,
         default:
             break
         }
-    }
-
-    // When app becomes active, and channel is open, recreate the database observers and reload
-    // the data source so that any missed database updates from the NotificationService are refreshed.
-    @objc func appMovedToForeground() {
-        messageController.delegate = self
-        messageListVC.dataSource = self
     }
 
     private func updateMessages(with changes: [ListChange<ChatMessage>]) {